Fixed circular dependency warnings. Updated layout and tab styling
