+++ packages/altair-app/src/app/components/query-editor/query-editor.component.html
@@ -2,7 +2,7 @@
   class="main-view-tabs"
   [nzAnimated]="false"
   nzSize="small"
-  nzTabBarGutter="20"
+  nzTabBarGutter="16"
   [(nzSelectedIndex)]="selectedIndex"
 >
   <nz-tab [nzTitle]="'QUERY_TAB' | translate">
