+++ packages/altair-app/src/app/store/index.ts
@@ -1,7 +1,6 @@
 import { InjectionToken } from '@angular/core';
 import { combineReducers, Action, ActionReducer, ActionReducerMap, MetaReducer } from '@ngrx/store';
-import { compose } from '@ngrx/store';
-import { LocalStorageConfig, localStorageSync } from 'ngrx-store-localstorage';
+import { localStorageSync } from 'ngrx-store-localstorage';
 import { storeFreeze } from 'ngrx-store-freeze';
 
 import { environment } from '../../environments/environment';
@@ -30,6 +29,7 @@ import { getAltairConfig } from 'app/config';
 import { TODO } from 'app/interfaces/shared';
 import { AppInitAction } from './action';
 import { asyncStorageSync } from './async-storage-sync';
+import { localStorageSyncConfig } from './local-storage-sync-config';
 
 export interface PerWindowState {
   layout: fromLayout.State;
@@ -85,19 +85,6 @@ export function log(_reducer: ActionReducer<any>): ActionReducer<any> {
     return _reducer(state, action);
   };
 }
-
-const getAltairInstanceStorageNamespace = () => getAltairConfig().initialData.instanceStorageNamespace || 'altair_';
-export const keySerializer = (key: string) => `${getAltairInstanceStorageNamespace()}${key}`;
-export const storageKeys = [ 'windows', 'windowsMeta', 'settings', 'environments' ];
-
-export const localStorageSyncConfig = {
-  keys: storageKeys,
-  rehydrate: true,
-  storage: performantLocalStorage,
-  restoreDates: false,
-  // syncCondition: (state) => console.log(state),
-  storageKeySerializer: keySerializer
-} as const;
 
 export function localStorageSyncReducer(_reducer: ActionReducer<any>): ActionReducer<any> {
   return localStorageSync(localStorageSyncConfig)(_reducer);
