+++ packages/altair-app/src/app/store/query/initialQuery.ts
@@ -1,9 +1,9 @@
 const initialQuery = `
-  # Welcome to Altair GraphQL Client.
-  # You can send your request using CmdOrCtrl + Enter.
+# Welcome to Altair GraphQL Client.
+# You can send your request using CmdOrCtrl + Enter.
 
-  # Enter your graphQL query here.
-`;
+# Enter your graphQL query here.
+`.trimStart();
 
 export {
   initialQuery
