+++ packages/altair-app/src/scss/_layout.scss
@@ -23,7 +23,7 @@
   flex: 1 1 auto;
   flex-direction: column;
   height: 100%;
-  padding: 10px;
+  padding: 5px;
   overflow-y: auto;
   -webkit-overflow-scrolling: touch;
 }
